requirements.txt
bootstrap-flask==2.2.0
Flask==2.2.5
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==9.5.0
numpy==1.24.3
matplotlib==3.7.1
gunicorn==20.1.0
requests==2.28.2